+++ crates/client/src/client.rs
@@ -12,6 +12,7 @@ use crate::client::{
         TransactionId,
     },
     types::{
+        gas_price::LatestGasPrice,
         message::MessageStatus,
         primitives::{
             Address,
@@ -346,6 +347,11 @@ impl FuelClient {
     pub async fn node_info(&self) -> io::Result<types::NodeInfo> {
         let query = schema::node_info::QueryNodeInfo::build(());
         self.query(query).await.map(|r| r.node_info.into())
+    }
+
+    pub async fn latest_gas_price(&self) -> io::Result<LatestGasPrice> {
+        let query = schema::gas_price::QueryLatestGasPrice::build(());
+        self.query(query).await.map(|r| r.latest_gas_price.into())
     }
 
     pub async fn connected_peers_info(&self) -> io::Result<Vec<PeerInfo>> {